requirements.txt
appnope==0.1.0
asgiref==3.3.1
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cloudinary==1.23.0
confusable-homoglyphs==3.2.0
cryptography==3.2.1
defusedxml==0.6.0
Django==3.1.3
django-admin-menu==1.4
django-allauth==0.43.0
django-better-admin-arrayfield==1.4.0
django-cors-headers==3.5.0
django-filter==2.4.0
django-registration==3.1.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
libsass==0.20.1
oauthlib==3.1.0
psycopg2-binary==2.8.6
gunicorn==20.0.4
pycparser==2.20
PyJWT==1.7.1
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.26.2
sentry-sdk==0.19.4
ddtrace==0.44.0
dj-database-url==0.5.0
whitenoise==5.2.0
django-debug-toolbar==3.2
django-letsencrypt==4.0.0
django-admin-menu==1.4
django-rest-registration==0.5.6